examples/repository_incident_analyzer/pyproject.toml
[build-system]
requires = ["setuptools >= 64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "repository-incident-analyzer"
version = "1.0.0"
description = "Combined Repository Analyzer (GitLab/GitHub) and PagerDuty Incident Analyzer using AIQ Toolkit"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "requests",
    "pydantic",
    "httpx",
    "python-dotenv",
]
requires-python = ">=3.11"

[project.entry-points."aiq.components"]
# Register combined tools with AIQ toolkit
repository_incident_tools = "repository_incident_analyzer.register"

[tool.setuptools.packages.find]
where = ["src"]
include = ["repository_incident_analyzer*"]

[tool.setuptools]
include-package-data = true 